requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.5.0
python-telegram-bot==22.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
